luming-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
asyncpg==0.29.0

# Redis
redis==5.0.1

# 数据源
akshare>=1.18.0
tushare>=1.4.0
httpx==0.26.0

# 定时任务
celery==5.3.6

# 工具
python-dotenv==1.0.0
apscheduler==3.10.4
